Move secure into its own folder
