+++ rai/lib/work.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <boost/optional.hpp>
-#include <rai/config.hpp>
+#include <rai/secure/config.hpp>
 #include <rai/lib/numbers.hpp>
 #include <rai/lib/utility.hpp>
