+++ rai/node/bootstrap.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include <rai/blockstore.hpp>
-#include <rai/ledger.hpp>
+#include <rai/secure/blockstore.hpp>
+#include <rai/secure/ledger.hpp>
 #include <rai/node/common.hpp>
 
 #include <atomic>
