+++ rai/node/rpc.hpp
@@ -5,7 +5,7 @@
 #include <boost/beast.hpp>
 #include <boost/property_tree/json_parser.hpp>
 #include <boost/property_tree/ptree.hpp>
-#include <rai/node/utility.hpp>
+#include <rai/secure/utility.hpp>
 #include <unordered_map>
 
 namespace rai
