+++ rai/secure/common.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <rai/lib/blocks.hpp>
-#include <rai/node/utility.hpp>
+#include <rai/secure/utility.hpp>
 
 #include <boost/property_tree/ptree.hpp>
