+++ rai/secure/icon.hpp
