+++ rai/secure/plat/osx/working.mm
