+++ rai/secure/plat/posix/working.cpp
