+++ rai/secure/plat/windows/working.cpp
