+++ rai/secure/versioning.cpp
@@ -1,4 +1,4 @@
-#include <rai/versioning.hpp>
+#include <rai/secure/versioning.hpp>
 
 rai::account_info_v1::account_info_v1 () :
 head (0),
