+++ rai/secure/versioning.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <rai/lib/blocks.hpp>
-#include <rai/node/utility.hpp>
+#include <rai/secure/utility.hpp>
 
 namespace rai
 {
